# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Whatnow API",
        "description": "Whatnow API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.whatnow.jazusoft.com/v1",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/subnationals": {
            "post": {
                "tags": [
                    "Regions"
                ],
                "summary": "Create a new subnational",
                "operationId": "createRegion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "countryCode",
                                    "title"
                                ],
                                "properties": {
                                    "countryCode": {
                                        "description": "Country code (3 characters)",
                                        "type": "string",
                                        "example": "USA"
                                    },
                                    "title": {
                                        "description": "Title of the subnational",
                                        "type": "string",
                                        "example": "North America"
                                    },
                                    "slug": {
                                        "description": "Slug for the subnational (optional)",
                                        "type": "string",
                                        "example": "north-america"
                                    },
                                    "translations": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "webUrl": {
                                                    "description": "Web URL for the translation",
                                                    "type": "string",
                                                    "format": "url",
                                                    "example": "https://example.com"
                                                },
                                                "lang": {
                                                    "description": "Language code (2 characters)",
                                                    "type": "string",
                                                    "example": "en"
                                                },
                                                "title": {
                                                    "description": "Title in the specified language",
                                                    "type": "string",
                                                    "example": "North America"
                                                },
                                                "description": {
                                                    "description": "Description in the specified language",
                                                    "type": "string",
                                                    "example": "Description of the subnational"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "deprecated": true,
                "security": []
            }
        },
        "/subnationals/subnational/{regionId}": {
            "put": {
                "tags": [
                    "Regions"
                ],
                "summary": "Update an existing subnational",
                "operationId": "updateRegion",
                "parameters": [
                    {
                        "name": "regionId",
                        "in": "path",
                        "description": "ID of the subnational to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Updated title of the subnational (optional)",
                                        "type": "string",
                                        "example": "Updated Region Title"
                                    },
                                    "slug": {
                                        "description": "Updated slug for the subnational (optional)",
                                        "type": "string",
                                        "example": "updated-subnational-slug"
                                    },
                                    "translations": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "webUrl": {
                                                    "description": "Web URL for the translation",
                                                    "type": "string",
                                                    "format": "url",
                                                    "example": "https://example.com"
                                                },
                                                "lang": {
                                                    "description": "Language code (2 characters)",
                                                    "type": "string",
                                                    "example": "en"
                                                },
                                                "title": {
                                                    "description": "Title in the specified language",
                                                    "type": "string",
                                                    "example": "Updated Title in Language"
                                                },
                                                "description": {
                                                    "description": "Description in the specified language",
                                                    "type": "string",
                                                    "example": "Updated description in language"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "deprecated": true,
                "security": []
            },
            "delete": {
                "tags": [
                    "Regions"
                ],
                "summary": "Delete a subnational",
                "operationId": "deleteRegion",
                "parameters": [
                    {
                        "name": "regionId",
                        "in": "path",
                        "description": "ID of the subnational to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "deprecated": true,
                "security": []
            }
        },
        "/subnationals/{country_code}": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Get all subnationals for a specific organisation by country code",
                "operationId": "d2296fe4eb4c2d71309b7bd2370494bd",
                "parameters": [
                    {
                        "name": "country_code",
                        "in": "path",
                        "description": "Country code of the organisation",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "deprecated": true,
                "security": []
            }
        },
        "/subnationals/{country_code}/{code}": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Get subnationals for a specific organisation by country code and language code",
                "operationId": "e80be67aaefaf2057958c0e54908fbc9",
                "parameters": [
                    {
                        "name": "country_code",
                        "in": "path",
                        "description": "Country code of the organisation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Language code for translations",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "deprecated": true,
                "security": []
            }
        },
        "/subnationals/subnational/translation/{translationId}": {
            "delete": {
                "tags": [
                    "Regions"
                ],
                "summary": "Delete a subnational translation",
                "operationId": "deleteTranslation",
                "parameters": [
                    {
                        "name": "translationId",
                        "in": "path",
                        "description": "ID of the translation to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "deprecated": true,
                "security": []
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "x-api-key"
            }
        }
    },
    "security": [
        {
            "BasicAuth": []
        }
    ]
}